Consolidated the insert/update steps for simplicity.
Using this method it should no longer require NewDayReceiver as we’re
always storing the most up to date steps for the current day at the
time of update.

Unless of course there’s another use for NewDayReceiver I’m not seeing
(I’ve not had a good look over the code).

The ShutdownRecevier is also VERY unreliable, what if the phone
crashes? I’m going to look at a better way of storing steps in the db
as I’ve lost steps myself from the app/phone crashing a few times.