Introduce getDaysWithoutToday for achievements
Create getDaysWithoutToday() to match together with
getTotalWithoutToday() and use it in the getDays() method.
getDaysWithoutToday() will return a value >= 0, getDays() will add 1 day
to it, so it will return a value >= 1.
Using the getDaysWithoutToday method for the achievement there is
already a check if days >= 10, so it is safe to divide by this value.